FIX: Modificando tokenJwt para ser enviado no Cookie ao em vez do body

+++ src/main/java/com/example/amigo_de_patas/dto/response/AuthResponse.java
@@ -1,11 +1,9 @@
 package com.example.amigo_de_patas.dto.response;
 
-import com.example.amigo_de_patas.model.Role;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.UUID;
 
 @Data
 @NoArgsConstructor
@@ -13,6 +11,6 @@ import java.util.UUID;
 public class AuthResponse {
     private String accessToken;
     private String refreshToken;
-    private Long expiresIn;
+    private Integer expiresIn;
     private String tokenType;
 }
