+++ src/main/java/com/example/amigo_de_patas/service/JwtService.java
@@ -18,11 +18,11 @@ public class JwtService {
 
     @Value("${jwt.expiration-in-seconds}")
     @Getter
-    private Long expirationInSeconds;
+    private Integer expirationInSeconds;
 
     @Value("${jwt.refresh-expiration-in-seconds}")
     @Getter
-    private Long refreshExpirationInSeconds;
+    private Integer refreshExpirationInSeconds;
 
     public JwtService(JwtEncoder jwtEncoder, JwtDecoder jwtDecoder) {
         this.jwtEncoder = jwtEncoder;
